requirements.txt
matplotlib==3.8.2
Pillow==10.1.0
